Move dbpedia query into utility
The dbpedia query is not the essense of the component for the
unindulged reader.  There's a whole lot to say to keep the function
inside of the component, I've extracted it to see if it looks easier
to understand.  This is dubious.
