+++ addon/components/editor-plugins/dbpedia-info-card.js
@@ -2,6 +2,7 @@ import { reads } from '@ember/object/computed';
 import Component from '@ember/component';
 import layout from '../../templates/components/editor-plugins/dbpedia-info-card';
 import fetch from 'fetch';
+import dbpediaQuery from '../../utils/editor-plugins/dbpedia-query';
 
 /**
 * Card displaying a hint of the Date plugin
@@ -27,34 +28,7 @@ export default Component.extend({
   // willRender method will get executed just before the card appears on the
   // screen, we use this method to fetch the information needed from dbpedia
   async willRender() {
-    const url = new URL("http://dbpedia.org/sparql");
-    const query = `
-      SELECT ?description ?image WHERE {
-        ?s rdfs:label "${this.info.term}"@en.
-        OPTIONAL {
-          ?s <http://www.w3.org/2000/01/rdf-schema#comment> ?description.
-          FILTER (lang(?description) = 'en')
-        }
-        OPTIONAL {
-          ?s <http://dbpedia.org/ontology/thumbnail> ?image.
-        }
-      }
-    `;
-    const params = {
-      format: "application/sparql-results+json",
-      query,
-    };
-    Object.keys(params).forEach(key => url.searchParams.append(key, params[key]));
-    const response = await fetch(url);
-    const json = await response.json();
-    const sparqlResult = json.results.bindings[0];
-
-    if (sparqlResult.description) {
-      this.set('description', sparqlResult.description.value);
-    }
-
-    if (sparqlResult.image) {
-      this.set('image', sparqlResult.image.value);
-    }
+    const foundInfo = await dbpediaQuery( this.info.term );
+    this.setProperties( foundInfo );
   },
 });
