+++ app/utils/editor-plugins/dbpedia-query.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-rdfa-editor-dbpedia-info-plugin/utils/editor-plugins/dbpedia-query';
