+++ tests/unit/utils/editor-plugins/dbpedia-query-test.js
@@ -1,0 +1,11 @@
+import editorPluginsDbpediaQuery from 'dummy/utils/editor-plugins/dbpedia-query';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | editor-plugins/dbpedia-query', function() {
+
+  // Replace this with your real tests.
+  test('it works', function(assert) {
+    let result = editorPluginsDbpediaQuery();
+    assert.ok(result);
+  });
+});
